istgt: Don't sleep for a full second after creating a worker
Otherwise inbound connections can linger in the listening socket's
accept queue for long enough that the initiator times out the
connection.  It's not a good idea to sleep in the main accept loop.

It's not obvious exactly why we want the new worker thread to start
running before we accept another connection, but we certainly don't have
to wait a full second.  Poll the connection state and sleep for 1ms
between polls instead.