istgt: Avoid sleeping unnecessarily while holding the connection mutex
When handling a iscsi login, istgt drops old connections to the same
LUN.  In so doing, it sleeps for one second while holding a global lock
that is acquired by the thread which accepts connections from
initiators.  Thus, this behaviour prevents the target from accepting
multiple initiator connections in quick succession.

In the common case this function has nothing to do, and the sleep is
pointless.  Avoid sleeping at all in that common case.